fixed ina219 problems, so now sweeping works

+++ task.c
@@ -3,8 +3,8 @@
 void mppt_task() { mppt_update(&mppt, pv_voltage_mv, pv_current_ma); }
 
 void boost_converter_sensor_task() {
-  pv_voltage_mv = read_voltage_mV(&boost_converter_ina);
-  pv_current_ma = read_current_mA(&boost_converter_ina);
+  pv_voltage_mv = (uint32_t)read_voltage_mV(&boost_converter_ina);
+  pv_current_ma = (uint32_t)read_current_mA(&boost_converter_ina);
 }
 
 void power_mode_stm_task() {
