+++ task.h
@@ -16,10 +16,10 @@ uint16_t bootcount_24hour;
 uint8_t exit_SAFE_1;
 
 // used by mppt, ina needs to update this value regularly
-uint16_t pv_voltage_mv;
+uint32_t pv_voltage_mv;
 
 // used by mppt , ina needs to update this value regularly
-uint16_t pv_current_ma;
+uint32_t pv_current_ma;
 
 stm_t power_mode_stm;
 ina219_t boost_converter_ina;
